05_Output/Projects/@Active/Kyoritsu-Seiyaku-DX/demo-data/requirements.txt
streamlit>=1.30.0
plotly>=5.18.0
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
snowflake-connector-python[pandas]>=3.6.0
